pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-gemini-rag"
version = "0.1.0"
description = "Self-correcting Agentic RAG system with LangGraph, Gemini, and Hybrid Search"
readme = "README.md"
requires-python = "==3.13.9"
license = {text = "MIT"}
authors = [
    {name = "Saeed Mehrang"}
]
keywords = ["rag", "langgraph", "gemini", "langchain", "hybrid-search", "faiss", "bm25"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "langchain==1.0.7",
    "langchain-community==0.4.1",
    "langchain-google-genai==3.0.3",
    "langgraph==1.0.3",
    "faiss-cpu==1.13.0",
    "pypdf==6.3.0",
    "unstructured==0.18.20",
    "python-dotenv==1.2.1",
    "rank-bm25==0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.1",
    "pytest-cov==7.0.0",
    "ruff==0.14.5",
    "mypy==1.18.2",
    "ipython==9.7.0",
    "ipykernel==7.1.0",
]

[project.scripts]
agentic-rag = "src.main:main"
index-docs = "scripts.index_documents:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "scripts*"]

[tool.ruff]
line-length = 100
target-version = "py313"
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort (Ruff handles import sorting)
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "B008", # Do not perform function calls in argument defaults (Good practice rule)
    "C901", # too complex (Good for educational/draft code)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true